bump to 3.4.0-M5, move agent to application package
