+++ pom.xml
@@ -4,7 +4,7 @@
   <parent>
     <groupId>io.akka</groupId>
     <artifactId>akka-javasdk-parent</artifactId>
-    <version>3.4.0-M1</version>
+    <version>3.4.0-M5</version>
   </parent>
   
   <groupId>org.example</groupId>
